Switched sprites to be stored at the first half of VPD tile memory

+++ project/source/engine/animation_types.h
@@ -72,7 +72,7 @@ typedef struct
     u8 numFrames;
     u8 tilesPerFrame;
     u8 totalTileCount;
-    u8* vdpLocation;
+    u16* vdpLocation;
 } TileAnimation;
